JAMES-3534 Split Identity and IdentityGet
