+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/mail/Identity.scala
@@ -23,24 +23,14 @@ import java.nio.charset.StandardCharsets
 
 import com.google.common.collect.ImmutableList
 import com.google.common.hash.Hashing
-import eu.timepit.refined.auto._
 import eu.timepit.refined.refineV
 import javax.inject.Inject
 import org.apache.james.core.MailAddress
 import org.apache.james.jmap.core.Id.Id
-import org.apache.james.jmap.core.UuidState.INSTANCE
-import org.apache.james.jmap.core.{AccountId, Properties, UuidState}
-import org.apache.james.jmap.method.WithAccountId
 import org.apache.james.mailbox.MailboxSession
 import org.apache.james.rrt.api.CanSendFrom
 
 import scala.jdk.CollectionConverters._
-
-object Identity {
-  val allProperties: Properties = Properties("id", "name", "email", "replyTo", "bcc", "textSignature", "htmlSignature",
-    "mayDelete")
-  val idProperty: Properties = Properties("id")
-}
 
 case class IdentityName(name: String) extends AnyVal
 case class TextSignature(name: String) extends AnyVal
@@ -59,28 +49,6 @@ case class Identity(id: IdentityId,
                     textSignature: Option[TextSignature],
                     htmlSignature: Option[HtmlSignature],
                     mayDelete: MayDeleteIdentity)
-
-case class IdentityGetRequest(accountId: AccountId,
-                              ids: Option[IdentityIds],
-                              properties: Option[Properties]) extends WithAccountId {
-  def computeResponse(identities: List[Identity]): IdentityGetResponse = {
-    val list: Option[List[Identity]] = Some(identities.filter(identity => isRequested(identity.id))).filter(_.nonEmpty)
-    val notFound: Option[List[IdentityId]] = ids.map(ids => ids.ids.filter(id => !identities.map(_.id).contains(id))).filter(_.nonEmpty)
-
-    IdentityGetResponse(
-      accountId = accountId,
-      state = INSTANCE,
-      list = list,
-      notFound = notFound.map(IdentityIds))
-  }
-
-  private def isRequested(id: IdentityId): Boolean = ids.forall(_.contains(id))
-}
-
-case class IdentityGetResponse(accountId: AccountId,
-                               state: UuidState,
-                               list: Option[List[Identity]],
-                               notFound: Option[IdentityIds])
 
 class IdentityFactory @Inject()(canSendFrom: CanSendFrom) {
   def listIdentities(session: MailboxSession): List[Identity] =
