+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/mail/IdentityGet.scala
@@ -1,0 +1,53 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.jmap.mail
+
+import eu.timepit.refined.auto._
+import org.apache.james.jmap.core.UuidState.INSTANCE
+import org.apache.james.jmap.core.{AccountId, Properties, UuidState}
+import org.apache.james.jmap.method.WithAccountId
+
+object IdentityGet {
+  val allProperties: Properties = Properties("id", "name", "email", "replyTo", "bcc", "textSignature", "htmlSignature",
+    "mayDelete")
+  val idProperty: Properties = Properties("id")
+}
+
+case class IdentityGetRequest(accountId: AccountId,
+                              ids: Option[IdentityIds],
+                              properties: Option[Properties]) extends WithAccountId {
+  def computeResponse(identities: List[Identity]): IdentityGetResponse = {
+    val list: Option[List[Identity]] = Some(identities.filter(identity => isRequested(identity.id))).filter(_.nonEmpty)
+    val notFound: Option[List[IdentityId]] = ids.map(ids => ids.ids.filter(id => !identities.map(_.id).contains(id))).filter(_.nonEmpty)
+
+    IdentityGetResponse(
+      accountId = accountId,
+      state = INSTANCE,
+      list = list,
+      notFound = notFound.map(IdentityIds))
+  }
+
+  private def isRequested(id: IdentityId): Boolean = ids.forall(_.contains(id))
+}
+
+case class IdentityGetResponse(accountId: AccountId,
+                               state: UuidState,
+                               list: Option[List[Identity]],
+                               notFound: Option[IdentityIds])
